added producer_id to chatroom table

+++ db/migrate/20200707111504_add_producer_id_to_chatroom.rb
@@ -1,5 +1,5 @@
 class AddProducerIdToChatroom < ActiveRecord::Migration[6.0]
   def change
-    add_column :chatrooms, :producer_id, :integer
+    add_column :chatrooms, :producer_id, :bigint
   end
 end
